refactor: extract Multimap
