+++ packages/client/test/test-utils/fake/FakeStreamRegistry.ts
@@ -13,7 +13,7 @@ import { BrubeckContainer } from '../../../src/Container'
 import { Ethereum } from '../../../src/Ethereum'
 import { StreamRegistry } from '../../../src/StreamRegistry'
 import { NotFoundError, SearchStreamsPermissionFilter } from '../../../src'
-import { Multimap } from '../utils'
+import { Multimap } from '../Multimap'
 import { StreamRegistryCached } from '../../../src/StreamRegistryCached'
 import { DOCKER_DEV_STORAGE_NODE } from '../../../src/ConfigTest'
 import { formStorageNodeAssignmentStreamId } from '../../../src/utils/utils'
